refactor : post/application 폴더로 이동
