+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindNonAuthPostListUsecase.java
@@ -5,7 +5,7 @@ import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import com.kakaotech.team14backend.outer.post.dto.GetHomePostListResponseDTO;
 import com.kakaotech.team14backend.outer.post.dto.SetNonAuthenticatedHomePostDTO;
-import com.kakaotech.team14backend.post.mapper.PostMapper;
+import com.kakaotech.team14backend.post.application.PostMapper;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
