+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindPostUsecase.java
@@ -4,7 +4,7 @@ import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import com.kakaotech.team14backend.outer.post.dto.GetPostDTO;
 import com.kakaotech.team14backend.outer.post.dto.GetPostResponseDTO;
-import com.kakaotech.team14backend.post.mapper.PostMapper;
+import com.kakaotech.team14backend.post.application.PostMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
